projects/ActionsUptime/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.44
botocore==1.35.90
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==44.0.2
defusedxml==0.7.1
dj-stripe==2.8.4
Django==5.1.7
django-allauth==65.7.0
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-redis==5.4.0
django-ses==4.2.0
django-timezone-field==7.1
gevent==24.2.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
phonenumbers==9.0.0
prettyprint==0.1.5
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-whois==0.9.5
python3-openid==3.2.0
redis==5.0.8
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.10.2
sentry-sdk==2.19.2
setuptools==80.7.1
six==1.16.0
sqlparse==0.5.3
stripe==4.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
